backend/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0
django-cleanup==8.1.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-ipware==7.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.7
ipapi==1.0.4
Pillow==10.1.0
PyJWT==2.8.0
python-ipware==3.0.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
stripe==9.7.0
typing_extensions==4.11.0
tzdata==2023.4
urllib3==2.2.1
gunicorn